Updated Files

+++ TableReader.java
@@ -1,4 +1,3 @@
-import java.sql.*;
 import java.util.concurrent.*;
 
 public class TableReader
@@ -11,7 +10,10 @@ public class TableReader
 
         sm.createTable();
 
-        //sm.insertData(new byte[1], 10);
+        if(args.length > 0 && args[0].equals("test"))
+        {
+            sm.insertData(new byte[1], 10);
+        }
 
         ConcurrentHashMap<String, FileData> fd = sm.selectAllFiles();
 
@@ -22,9 +24,13 @@ public class TableReader
 
         else
         {
-            fd.forEach((k, v) -> System.out.printf("Name: %s\nSize:%d\nData: %s", v.fileName, v.fileSize, new String(v.data, 0, v.data.length)));
+            fd.forEach((k, v) -> System.out.printf("Name: %s\nSize:%d\nData: %s\n", v.fileName, v.fileSize, new String(v.data, 0, v.data.length)));
         }
 
-        sm.dropTable();
+
+        if(args.length > 0 && args[0].equals("test"))
+        {
+            sm.dropTable();
+        }
     }
 }
